2024.11.17 项目主体代码结束，测试用例上传

+++ Base/Detail/config.hpp
@@ -41,7 +41,7 @@ constexpr short MAX_SHORT = 32767;
 
 constexpr short MIN_SHORT = -32768;
 
-constexpr short MAX_USHORT = -1;
+constexpr unsigned short MAX_USHORT = static_cast<unsigned short>(-1);
 
 constexpr int MAX_INT = 2147483647;
